Hey there past self, this is your future self speaking. We need to talk.
I know you think that the local changes you've done so far are too messy
to commit as-is. I know you feel it's very important to break it up into
a logical, well-commented commit sequence with issue references and all.

But look at this shit, just fucking look at it. Do you think that if you
kick the can down the road you'll actually remember how this pile of hot
garbage was supposed to be split into smaller changes? Because it's been
exactly two months now and I sure as fuck don't have a clue anymore.

You should be thankful that going backwards in time is likely impossible
because otherwise I'd come and slap you so hard that doing early commits
won't feel painful at all in comparison.